+++ app/components/about/index.jsx
@@ -1,20 +1,20 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import { Link } from 'react-router';
 import { Grid } from 'react-bootstrap';
+import { translate } from 'react-i18next';
 import { paths } from 'helpers/routes';
 
-const About = ({ children }) => {
+const About = ({ children, t }) => {
   const id = 'test_id';
 
   return (
     <Grid>
       <article>
-        <h1>About</h1>
-        <p>
-          Kick-start your new web application based on React and Redux technologies.
-        </p>
+        <h1>{ t('about:title') }</h1>
+        <p>{ t('about:text') }</p>
         <Link to={ paths.aboutExtended(id) }>
-          show details...
+          { t('about:details') }
         </Link>
         { children }
       </article>
@@ -22,4 +22,8 @@ const About = ({ children }) => {
   );
 };
 
-export default About;
+About.propTypes = {
+  t: PropTypes.func.isRequired
+};
+
+export default translate()(About);