+++ app/containers/signinModal/index.jsx
@@ -1,18 +1,19 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
+import { translate } from 'react-i18next';
 import { Modal } from 'react-bootstrap';
 import sessionActions from 'actions/session';
 import SigninForm from 'components/signinForm';
 
-const SigninModalContainer = ({ isOpen, isLoading, signinUser, closeModal }) => (
+const SigninModalContainer = ({ isOpen, isLoading, signinUser, closeModal, t }) => (
   <Modal
     bsSize="small"
     show={ isOpen }
     onHide={ closeModal }
   >
     <Modal.Header closeButton>
-      <h3 className="modal-title">Sign In</h3>
+      <h3 className="modal-title">{ t('modal:signin') }</h3>
     </Modal.Header>
     <SigninForm
       isLoading={ isLoading }
@@ -26,7 +27,8 @@ SigninModalContainer.propTypes = {
   closeModal: PropTypes.func.isRequired,
   isLoading: PropTypes.bool.isRequired,
   isOpen: PropTypes.bool.isRequired,
-  signinUser: PropTypes.func.isRequired
+  signinUser: PropTypes.func.isRequired,
+  t: PropTypes.func.isRequired
 };
 
 const mapStateToProps = state => ({
@@ -37,4 +39,4 @@ const mapDispatchToProps = dispatch => ({
   signinUser: user => dispatch(sessionActions.signinUser(user))
 });
 
-export default connect(mapStateToProps, mapDispatchToProps)(SigninModalContainer);
+export default connect(mapStateToProps, mapDispatchToProps)(translate()(SigninModalContainer));